packages/pyapi/pyproject.toml
[tool.poetry]
name = "pyapi"
version = "0.1.0"
description = "Whisper-tiny + CLIP + Mistral-7B-Instruct (CPU)"
authors = ["Camilla Thorsen <camilla@thorsen.dev>"]
readme = "README.md"
packages = [{ include = "api_app" }]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
transformers = "^4.40.0"
sentencepiece = "^0.2.0"
pillow = "^10.3.0"
llama-cpp-python = { version = "^0.2.72", extras = ["server"] }
torch = "2.2.2"
numpy = "^1.26"
python-multipart = "^0.0.9"
pydub = "^0.25.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
